web-gui/config.py
from dotenv import load_dotenv
import os

load_dotenv()

class Config:
    DB_HOST = os.getenv('DB_HOST', 'localhost')
    DB_PORT = int(os.getenv('DB_PORT', 3306))
    DB_NAME = os.getenv('DB_NAME', 'goldbitna_l200')
    DB_USER = os.getenv('DB_USER', 'root')
    DB_PASSWORD = os.getenv('DB_PASSWORD', '')
    
    SERVER_PORT = int(os.getenv('SERVER_PORT', 8080))
    DEBUG = os.getenv('DEBUG', 'true').lower() == 'true'