web-gui/requirements.txt
nicegui==2.10.0
fastapi
uvicorn[standard]
mysql-connector-python
python-dotenv
pydantic
jinja2